"Improve documentation by fixing grammatical errors and updating outdated information"
This commit focuses on refining the document on setting up Kubernetes. It corrects typographical mistakes like missing commas, adopts the correct spelling for "Wi-Fi". This goes a long way in ensuring readers comprehend the instructions more effectively. The commit also includes updated guidelines for the generation of SSH key pairs and configuring Raspberry Pi for Wi-Fi connection. This makes the document current and more useful to those implementing the instructions.